Changed GetMessagesForUser was changed to post
-We changed it to a post cause we needed to send the messageParams to
the controller. We could have done it through the parameters but this is
the way yewo decided to do it. I will attempt to do the same with
GetThread so maybe we can use the parameter, but the errors I was
getting were because not only was I not passing in any info but also it
was not the right info